Funnel patches and new posts to same Transaction route
This allows us to reuse the code which adds a message
to a transaction, and also simplifies some of the client
side code. The idea now is that if a transaction id
(t_id) is included in the body of the post, it is
assumed to be an update to an existing transaction.
Otherwise, it is assumed to be posting a new transaction.